requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pytest==7.4.0
pytest-mock==3.11.1
black==23.7.0
flake8==6.0.0
mypy==1.4.1
isort==5.12.0